UNIAPI/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
argon2-cffi==23.1.0
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
itsdangerous==2.2.0
starlette==0.38.5
aiosqlite==0.20.0     # necessário para sqlite+aiosqlite

# ---- opcionais para outros bancos (comente/descomente conforme uso) ----
# asyncpg==0.29.0      # PostgreSQL (postgresql+asyncpg://)
# psycopg[binary]==3.1.19  # alternativa Postgres síncrona
# oracledb==2.2.1      # Oracle (modo thin)
# pyodbc==5.1.0        # SQL Server via ODBC (Windows)
